SNAP-159: Production people summary out of memory
