+++ src/main/java/gov/ca/cwds/neutron/util/jdbc/NeutronDB2Utils.java
@@ -5,8 +5,6 @@ import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
 import java.util.Date;
 
-import org.apache.commons.lang3.StringUtils;
-
 import com.ibm.db2.jcc.DB2Connection;
 import com.ibm.db2.jcc.DB2SystemMonitor;
 
@@ -24,27 +22,10 @@ import gov.ca.cwds.neutron.util.shrinkray.NeutronDateUtils;
  */
 public final class NeutronDB2Utils {
 
-  private static final ConditionalLogger LOGGER = new JetPackLogger(NeutronDB2Utils.class);
+  static final ConditionalLogger LOGGER = new JetPackLogger(NeutronDB2Utils.class);
 
   private NeutronDB2Utils() {
     // Default no-op.
-  }
-
-  /**
-   * Enable DB2 parallelism. Ignored for other databases.
-   * 
-   * @param con connection
-   * @throws SQLException connection error
-   */
-  public static void enableBatchSettings(Connection con) throws SQLException {
-    final String dbProductName = con.getMetaData().getDatabaseProductName();
-    con.setSchema(NeutronJdbcUtils.getDBSchemaName());
-    con.setAutoCommit(false);
-
-    if (StringUtils.containsIgnoreCase(dbProductName, "db2")) {
-      LOGGER.info("Apply DB2 batch settings");
-      new WorkSetDB2UserInfo().execute(con);
-    }
   }
 
   /**
