+++ src/main/java/gov/ca/cwds/neutron/util/jdbc/WorkSecondaryResults.java
@@ -41,7 +41,7 @@ public class WorkSecondaryResults<T extends PersistentObject> extends NeutronWor
   @Override
   public void execute(Connection con) throws SQLException {
     setConnection(con);
-    NeutronDB2Utils.enableBatchSettings(con);
+    NeutronJdbcUtils.enableBatchSettings(con);
     handler.handleSecondaryJdbc(con, Pair.<String, String>of("a", "b"));
   }
